requirements.txt
pandas==2.2.2
numpy==2.1.0
loguru==0.7.2
tqdm==4.66.5
selenium==4.23.1
pyyaml==6.0.2